Add files via upload
Code implemented using the FLOWVlm simulator of a blown wing. Simulator uses dimension hard coded into the main, and then sets up a monitor function to iterate like the examples shown for FLOWUnsteady (heavingwing, etc.). The code also extracts the forces on the wing (drag and lift) and forces per unit span, plotting them on with each iteration to show forces over the duration of the maneuver.